+++ oshi-core/src/main/java/oshi/software/common/AbstractOperatingSystem.java
@@ -22,7 +22,6 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Comparator;
-import java.util.LinkedList;
 import java.util.List;
 
 import oshi.software.os.OSProcess;
@@ -180,8 +179,16 @@ public abstract class AbstractOperatingSystem implements OperatingSystem {
      * {@inheritDoc}
      */
     @Override
+    public OSProcess[] getProcesses(int limit, ProcessSort sort) {
+        return getProcesses(limit, sort, false);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
     public List<OSProcess> getProcesses(Collection<Integer> pids) {
-        List<OSProcess> returnValue = new LinkedList<>();
+        List<OSProcess> returnValue = new ArrayList<>(pids.size());
         for (Integer pid : pids) {
             OSProcess process = getProcess(pid);
             if (process != null) {