+++ oshi-core/src/main/java/oshi/software/os/linux/LinuxOperatingSystem.java
@@ -165,13 +165,13 @@ public class LinuxOperatingSystem extends AbstractOperatingSystem {
      * {@inheritDoc}
      */
     @Override
-    public OSProcess[] getProcesses(int limit, ProcessSort sort) {
+    public OSProcess[] getProcesses(int limit, ProcessSort sort, boolean slowFields) {
         List<OSProcess> procs = new ArrayList<>();
         File[] pids = ProcUtil.getPidFiles();
 
         // now for each file (with digit name) get process info
         for (File pid : pids) {
-            OSProcess proc = getProcess(ParseUtil.parseIntOrDefault(pid.getName(), 0));
+            OSProcess proc = getProcess(ParseUtil.parseIntOrDefault(pid.getName(), 0), slowFields);
             if (proc != null) {
                 procs.add(proc);
             }
@@ -185,6 +185,10 @@ public class LinuxOperatingSystem extends AbstractOperatingSystem {
      */
     @Override
     public OSProcess getProcess(int pid) {
+        return getProcess(pid, true);
+    }
+
+    private OSProcess getProcess(int pid, boolean slowFields) {
         String path = "";
         Pointer buf = new Memory(1024);
         int size = Libc.INSTANCE.readlink(String.format("/proc/%d/exe", pid), buf, 1023);
@@ -223,8 +227,10 @@ public class LinuxOperatingSystem extends AbstractOperatingSystem {
         proc.setBytesWritten(ParseUtil.parseLongOrDefault(MapUtil.getOrDefault(io, "write_bytes", ""), 0L));
 
         // gets the open files count
-        List<String> openFilesList = ExecutingCommand.runNative(String.format("ls -f /proc/%d/fd", pid));
-        proc.setOpenFiles(openFilesList.size() - 1L);
+        if (slowFields) {
+            List<String> openFilesList = ExecutingCommand.runNative(String.format("ls -f /proc/%d/fd", pid));
+            proc.setOpenFiles(openFilesList.size() - 1L);
+        }
 
         Map<String, String> status = FileUtil.getKeyValueMapFromFile(String.format("/proc/%d/status", pid), ":");
         proc.setName(MapUtil.getOrDefault(status, "Name", ""));
@@ -283,7 +289,7 @@ public class LinuxOperatingSystem extends AbstractOperatingSystem {
         for (File procFile : procFiles) {
             int pid = ParseUtil.parseIntOrDefault(procFile.getName(), 0);
             if (parentPid == getParentPidFromProcFile(pid)) {
-                OSProcess proc = getProcess(pid);
+                OSProcess proc = getProcess(pid, true);
                 if (proc != null) {
                     procs.add(proc);
                 }