+++ oshi-json/src/main/java/oshi/json/software/os/OperatingSystem.java
@@ -67,7 +67,7 @@ public interface OperatingSystem extends OshiJsonObject {
      * The order may be specified by the sort parameter, for example, to return
      * the top cpu or memory consuming processes; if null, no order is
      * guaranteed.
-     *
+     * 
      * In the JSON output, these parameters may be specified by
      * "operatingSystem.processes.limit" and "operatingSystem.processes.sort"
      * property values.
@@ -79,9 +79,38 @@ public interface OperatingSystem extends OshiJsonObject {
      * @return An array of {@link oshi.software.os.OSProcess} objects for the
      *         specified number (or all) of currently running processes, sorted
      *         as specified. The array may contain null elements if a process
-     *         terminates during iteration.
+     *         terminates during iteration. Some fields that are slow to
+     *         retrieve (e.g., group information on Windows, open files on Unix
+     *         and Linux) will be skipped.
      */
     OSProcess[] getProcesses(int limit, ProcessSort sort);
+
+    /**
+     * Gets currently running processes. If a positive limit is specified,
+     * returns only that number of processes; zero will return all processes.
+     * The order may be specified by the sort parameter, for example, to return
+     * the top cpu or memory consuming processes; if null, no order is
+     * guaranteed.
+     * 
+     * In the JSON output, these parameters may be specified by
+     * "operatingSystem.processes.limit" and "operatingSystem.processes.sort"
+     * property values.
+     *
+     * @param limit
+     *            Max number of results to return, or 0 to return all results
+     * @param sort
+     *            If not null, determines sorting of results
+     * @param slowFields
+     *            If false, skip {@link oshi.software.os.OSProcess} fields that
+     *            are slow to retrieve (e.g., group information on Windows, open
+     *            files on Unix and Linux). If true, include all fields,
+     *            regardless of how long it takes to retrieve the data.
+     * @return An array of {@link oshi.software.os.OSProcess} objects for the
+     *         specified number (or all) of currently running processes, sorted
+     *         as specified. The array may contain null elements if a process
+     *         terminates during iteration.
+     */
+    OSProcess[] getProcesses(int limit, ProcessSort sort, boolean slowFields);
 
     /**
      * Gets information on a currently running process
@@ -94,9 +123,9 @@ public interface OperatingSystem extends OshiJsonObject {
     OSProcess getProcess(int pid);
 
     /**
-     * Gets information on a currently running processes. This was primarily
-     * written to provide an optimized mechanism for windows based operating
-     * systems.
+     * Gets information on a currently running processes. This has improved
+     * performance on Windows based operating systems vs. iterating individual
+     * processes.
      *
      * @param pids
      *            A collection of process IDs